fix : Image.cpp/h

+++ Jazz Jack Rabbit 2/Image.cpp
@@ -119,7 +119,7 @@ void Image::Render(HDC hdc)
 		SRCCOPY);			// 복사 옵션
 }
 
-void Image::Render(HDC hdc, float destX, float destY)
+void Image::Render(HDC hdc, int destX, int destY)
 {
 	if (isTransparent)
 	{
@@ -150,7 +150,7 @@ void Image::Render(HDC hdc, float destX, float destY)
 
 }
 
-void Image::Render(HDC hdc, float destX, float destY, int frameX, int frameY, float scale)
+void Image::Render(HDC hdc, int destX, int destY, int frameX, int frameY, float scale)
 {
 	// frameX : 0, frameY : 0 => 시작 (68 * 0, 0) 얼마나 복사할건가 (68, 104)
 	// frameX : 1, frameY : 0 => 시작 (68 * 1, 0)  (68, 104)
