+++ Jazz Jack Rabbit 2/Image.h
@@ -62,8 +62,8 @@ public:
 	void Release();	// 메모리 해제
 
 	void Render(HDC hdc);
-	void Render(HDC hdc, float destX, float destY);	// 이미지 데이터를 화면에 복사
-	void Render(HDC hdc, float destX, float destY, int frameX, int frameY, float scale = 1.0f);	// 이미지 데이터를 화면에 복사
+	void Render(HDC hdc, int destX, int destY);	// 이미지 데이터를 화면에 복사
+	void Render(HDC hdc, int destX, int destY, int frameX, int frameY, float scale = 1.0f);	// 이미지 데이터를 화면에 복사
 
 	HDC GetMemDC() { if (imageInfo) return imageInfo->hMemDc; return NULL; }
